Important refactors
