+++ app/src/main/java/com/doers/games/geohangman/controllers/StartChallengeActivity.java
@@ -10,19 +10,13 @@ import android.os.Bundle;
 import android.os.Parcelable;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 import android.widget.ImageView;
-import android.widget.LinearLayout;
-import android.widget.Toast;
 
 import com.doers.games.geohangman.R;
 import com.doers.games.geohangman.constants.Constants;
 import com.doers.games.geohangman.constants.Messages;
-import com.doers.games.geohangman.custom_components.LetterButton;
-import com.doers.games.geohangman.custom_components.LetterButtonFactory;
 import com.doers.games.geohangman.model.Challenge;
 import com.doers.games.geohangman.services.IGeoHangmanService;
-import com.doers.games.geohangman.utils.ChallengeUtils;
 import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.SupportMapFragment;
@@ -33,6 +27,13 @@ import com.google.inject.Inject;
 import roboguice.activity.RoboActionBarActivity;
 import roboguice.inject.InjectView;
 
+/**
+ *
+ * This is the Start Challenge Activity. This activity starts when the opponent has received the challenge.
+ * Here the opponent has to guess the word.
+ *
+ * * @author <a href="mailto:aajn88@gmail.com">Antonio Jimenez</a>
+ */
 public class StartChallengeActivity extends RoboActionBarActivity {
 
     /** GeoHangman main service **/
@@ -43,15 +44,8 @@ public class StartChallengeActivity extends RoboActionBarActivity {
     @InjectView(R.id.challengePicIv)
     private ImageView mChallengePicIv;
 
-    /** LinearLayout where final word will be placed **/
-    @InjectView(R.id.finalWordLl)
-    private LinearLayout mFinalWordLl;
-
-    /** LinearLayout where disordered letters will be placed **/
-    @InjectView(R.id.lettersLl)
-    private LinearLayout mLettersLl;
-
-    private WordViewManager wordViewManager;
+    @Inject
+    private StartChallengeActivityHelper startChallengeHelper;
 
     /** Google Map **/
     private GoogleMap mMap;
@@ -114,7 +108,6 @@ public class StartChallengeActivity extends RoboActionBarActivity {
     private void setUpChallengeViews() {
         mChallengePicIv.setImageBitmap(geoHangmanService.getStoredPic());
         setUpMap();
-        wordViewManager = new WordViewManager(geoHangmanService.getStoredWord());
     }
 
     /**
@@ -131,119 +124,6 @@ public class StartChallengeActivity extends RoboActionBarActivity {
             LatLng latLng = new LatLng(point.getLat(), point.getLng());
             mMap.addMarker(new MarkerOptions().position(latLng));
             mMap.animateCamera(CameraUpdateFactory.newLatLngZoom(latLng, point.getZoom()));
-        }
-    }
-
-    private class WordViewManager {
-
-        /** Array with all LetterButton that the User has selected **/
-        private LetterButton[] finalButtons;
-
-        /** Array with all LetterButton that the User can select **/
-        private LetterButton[] lettersButtons;
-
-        /** Current Empty position **/
-        private Integer currentEmptyPos = 0;
-
-        /** Flag to search again the next empty pos **/
-        private Boolean searchAgain = Boolean.FALSE;
-
-        /** Letter counter: To know how many letters has been selected **/
-        private Integer letterCount = 0;
-
-        public WordViewManager(String word) {
-            setUpWord(word);
-        }
-
-        private void setUpWord(String word) {
-
-            String disorderedWord = ChallengeUtils.disorderString(word);
-
-            finalButtons = new LetterButton[disorderedWord.length()];
-            lettersButtons = new LetterButton[disorderedWord.length()];
-
-            mFinalWordLl.removeAllViews();
-            mLettersLl.removeAllViews();
-
-            for (int i = 0; i < disorderedWord.length(); i++) {
-                final LetterButton emptyLetter = LetterButtonFactory.createEmptyLetterButton(StartChallengeActivity.this);
-                finalButtons[i] = emptyLetter;
-                mFinalWordLl.addView(emptyLetter);
-                emptyLetter.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        removeLetter(emptyLetter);
-                    }
-                });
-
-                final LetterButton letterButton = LetterButtonFactory.
-                        createLetterButton(StartChallengeActivity.this, disorderedWord.charAt(i), i);
-                lettersButtons[i] = letterButton;
-                mLettersLl.addView(letterButton);
-                letterButton.setOnClickListener(new View.OnClickListener() {
-                    @Override
-                    public void onClick(View v) {
-                        addLetter(letterButton);
-                    }
-                });
-            }
-        }
-
-        private void addLetter(LetterButton letterButton) {
-            Integer index = nextEmptyPosition();
-            if(index == null) {
-                return;
-            }
-
-            finalButtons[index].setLetter(letterButton.getLetter(), letterButton.getIndex());
-            letterButton.setEmpty();
-            letterCount++;
-
-            if(letterCount == finalButtons.length) {
-                String message = "Wrong Word! Try again!";
-                if(geoHangmanService.verifyWord(buildFinalWord())) {
-                    message = "You got it! You win!";
-                }
-
-                Toast.makeText(StartChallengeActivity.this, message, Toast.LENGTH_LONG).show();
-            }
-        }
-
-        private String buildFinalWord() {
-            StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < finalButtons.length; i++) {
-                sb.append(finalButtons[i].getLetter());
-            }
-            return sb.toString();
-        }
-
-        private void removeLetter(LetterButton letterButton) {
-            lettersButtons[letterButton.getIndex()].setLetter(letterButton.getLetter(), letterButton.getIndex());
-            letterButton.setEmpty();
-            searchAgain = Boolean.TRUE;
-            letterCount--;
-        }
-
-        /**
-         * This method returns the next empty position
-         *
-         * @return next empty position
-         */
-        private Integer nextEmptyPosition() {
-            Integer pos = null;
-            if(searchAgain) {
-                Boolean found = Boolean.FALSE;
-                for (int i = 0; i < finalButtons.length && !found; i++) {
-                    if(finalButtons[i].isEmpty()) {
-                        pos = currentEmptyPos = i;
-                        found = true;
-                    }
-                }
-            } else {
-                pos = (currentEmptyPos >= finalButtons.length ? null : currentEmptyPos++);
-            }
-
-            return pos;
         }
     }
 
@@ -284,6 +164,7 @@ public class StartChallengeActivity extends RoboActionBarActivity {
         protected void onPostExecute(Void aVoid) {
             super.onPostExecute(aVoid);
             setUpChallengeViews();
+            startChallengeHelper.setUpWord();
             progress.dismiss();
         }
     }
