+++ app/src/main/java/com/doers/games/geohangman/utils/StoreUtils.java
@@ -1,14 +1,8 @@
 package com.doers.games.geohangman.utils;
 
-import android.content.Context;
-import android.graphics.Bitmap;
 import android.os.Environment;
 
-import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
-
 import java.io.File;
-import java.io.FileOutputStream;
 import java.io.IOException;
 
 /**
@@ -20,6 +14,9 @@ public final class StoreUtils {
 
     /** Constant for temporary folder **/
     public static final String TEMPORARY_FOLDER = "/.temp/";
+
+    /** Private no-parameters constructor **/
+    private StoreUtils(){}
 
     /**
      * This method creates a temporary file to store the pic
